refactor : 불필요 훅 삭제
