+++ src/1_Page/Profile/ui/PlayerDashBoard/util/convert.ts
@@ -1,10 +1,11 @@
 import { matchPosition } from "../../../../../4_Shared/constant/matchPosition";
 import { platform } from "../../../../../4_Shared/constant/platform";
 import { commonStatusIdx } from "../../../../../4_Shared/constant/commonStatusIdx";
-import { UserInfoInput } from "../type";
-import { UserInfoPost } from "../../../../../3_Entity/Account/type";
 
-export const converPostData = (data: UserInfoInput): UserInfoPost => {
+import { UserInfoForm, UserInfoProps } from "../type";
+import { UserInfoPost } from "../../../../../3_Entity/Account/types/request";
+
+export const convertToPostData = (data: UserInfoForm): UserInfoPost => {
   const postFormData = {
     state_message: data.state_message,
     nickname: data.nickname,
@@ -14,4 +15,13 @@ export const converPostData = (data: UserInfoInput): UserInfoPost => {
     common_status_idx: commonStatusIdx.indexOf(data.common_status_idx),
   };
   return postFormData;
+};
+
+export const convetToInfoForm = (userInfo: UserInfoProps): UserInfoForm => {
+  return {
+    ...userInfo,
+    platform: platform[Number(userInfo.platform)],
+    position: matchPosition[Number(userInfo.position)],
+    common_status_idx: commonStatusIdx[Number(userInfo.common_status_idx)],
+  };
 };
