+++ src/3_Entity/Account/types/request.d.ts
@@ -1,0 +1,7 @@
+export type UserInfoPost = {
+  nickname: string;
+  platform: number;
+  common_status_idx: number;
+  position: number;
+  state_message: string;
+};
