+++ src/3_Entity/Account/usePutUserInfo.ts
@@ -1,13 +1,10 @@
 import React from "react";
 import { useFetch } from "../../4_Shared/util/apiUtil";
-import { UserInfoPost } from "./type";
+import { UserInfoPost } from "./types/request";
 
-const usePutUserInfo = ({
-  userIdx,
-}: {
-  userIdx: number;
-  onFail?: () => void;
-}): [
+const usePutUserInfo = (
+  userIdx: number
+): [
   postEvent: (userInfo: UserInfoPost) => void,
   serverState: unknown,
   loading: boolean
@@ -23,7 +20,6 @@ const usePutUserInfo = ({
     if (!serverState) return;
     switch (serverState.status) {
       case "403":
-        onFail();
     }
   }, [serverState]);
